Move filter function declarations to utility headers
* Declarations were moved to utility functions
* Testperf tests were corrected because of declarations moving
* Includes of filter.h were removed (utility_float.h should be used for
  nm0)
