+++ src_proc0/common/textureTriangle_nm0.cpp
@@ -6,7 +6,6 @@
 #include <stdio.h>
 #include <math.h>
 #include <float.h> //TODO: only FLT_EPSILON is used from float.h
-#include "filter.h"
 #include "utility_float.h"
 
 #define TEXTURE_TRIANGLE_SECTION ".text"
