+++ test/perf/mai/Makefile
@@ -8,7 +8,7 @@ all_mc12101: configure0 configure1
 configure0 config_nmpu0:
 	testperf config -i ../../../include/nmgltex_nm0.h -b make_mc12101_nmpu0-ld
 	testperf config -i ../../../include/service.h -b make_mc12101_nmpu0-ld
-	testperf config -i ../../../include/filter.h -b make_mc12101_nmpu0-ld
+	testperf config -i ../../../utilities/include/utility_float.h -b make_mc12101_nmpu0-ld
 
 configure1 config_nmpu1:
 	testperf config -i ../../../include/nmgltex_nm1.h -b make_mc12101_nmpu1-ld --point fixed
